requirements.txt
PyYAML>=6.0
jsonschema>=4.17.0
requests>=2.28.0